chore(frontend): scaffold vue router
