+++ slice-guard/frontend/package.json
@@ -10,7 +10,8 @@
     "serve": "vite preview"
   },
   "dependencies": {
-    "vue": "^3.5.13"
+    "vue": "^3.5.13",
+    "vue-router": "^4.3.2"
   },
   "devDependencies": {
     "tailwindcss": "^3.4.4",
