+++ slice-guard/frontend/src/main.ts
@@ -1,5 +1,6 @@
 import { createApp } from 'vue'
 import './styles/main.css'
 import App from './App.vue'
+import router from './router'
 
-createApp(App).mount('#app')
+createApp(App).use(router).mount('#app')
